# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Food API"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Register",
                "description": "Register",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong Credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Get user data",
                "description": "Get user data",
                "operationId": "99e9c2c91427f6bf63c2f0ef765a768f",
                "responses": {
                    "200": {
                        "description": "User data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/fdc/food/search/{query}/{page}": {
            "get": {
                "tags": [
                    "FoodData Central"
                ],
                "summary": "Search for food",
                "operationId": "e0be62c9e7fbe7d4a5e77f95b20bb521",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of food products from FoodData Central"
                    }
                }
            }
        },
        "/fdc/food/{id}": {
            "get": {
                "tags": [
                    "FoodData Central"
                ],
                "summary": "Get food product",
                "operationId": "a1ae73a1f86e97bee27c0ddbcc523536",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}